Fixa json problem eric

+++ klient-projekttermin/src/com/example/klien_projekttermin/database/DatabaseHandlerAssignment.java
@@ -175,8 +175,8 @@ public class DatabaseHandlerAssignment extends SQLiteOpenHelper {
 				Assignment assignment = new Assignment(Long.valueOf(cursor
 						.getString(0)), // id från DB
 						cursor.getString(1), // name
-						Long.parseLong(cursor.getString(2)), // lat
-						Long.parseLong(cursor.getString(3)), // lon
+						Double.parseDouble(cursor.getString(2)), // lat
+						Double.parseDouble(cursor.getString(3)), // lon
 						cursor.getString(4),// region
 						agents, // agents
 						cursor.getString(6), // sender
